fix(01-01): change in_transaction from bool* to int* across C API and all bindings
- C API header and impl use int* out_active for ABI consistency
- C API test uses int variable with EQ/NE assertions
- Julia ccall uses Ptr{Cint}, wrapper returns Bool via != 0
- Dart binding uses ffi.Pointer<ffi.Int32>, wrapper returns bool via != 0
- Python cdef uses int*, wrapper uses ffi.new("int*")
- JS binding uses Int32Array(1) instead of Uint8Array(1)

+++ bindings/js/src/transaction.ts
@@ -29,7 +29,7 @@ Database.prototype.rollback = function (this: Database): void {
 
 Database.prototype.inTransaction = function (this: Database): boolean {
   const lib = getSymbols();
-  const outActive = new Uint8Array(1);
+  const outActive = new Int32Array(1);
   check(lib.quiver_database_in_transaction(this._handle, ptr(outActive)));
   return outActive[0] !== 0;
 };
